Share CIDConverter code between demos.
