+++ leshan-server-core-demo/pom.xml
@@ -34,6 +34,10 @@ Contributors:
             <artifactId>leshan-server-cf</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.eclipse.leshan</groupId>
+            <artifactId>leshan-core-demo</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
         </dependency>
