+++ leshan-server-core-demo/src/main/java/org/eclipse/leshan/server/core/demo/cli/converters/ServerCIDConverter.java
@@ -1,0 +1,9 @@
+package org.eclipse.leshan.server.core.demo.cli.converters;
+
+import org.eclipse.leshan.core.demo.cli.converters.CIDConverter;
+
+public class ServerCIDConverter extends CIDConverter {
+    public ServerCIDConverter() {
+        super(6);
+    }
+};
